requirements/common.txt
dj-database-url==0.4.1
django-configurations==2.0.0
django-debug-toolbar==1.6.0
django-dotenv==1.4.1
django-extensions==1.7.5
django-markdown-app
Django==1.10.3
psycopg2==2.6.2
pillow
